requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
slack_sdk==3.27.1
motor==3.3.2
dnspython==2.6.1
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
langdetect==1.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
itsdangerous==2.2.0
python-multipart==0.0.9
